Move store logic in is own file
